Fix typo in variable declaration

+++ bit_manipulation/3-set_bit.c
@@ -14,7 +14,7 @@
 int set_bit(unsigned long int *n, unsigned int index)
 {
     int mask = 1 << index;
-    int digit = *n;
+    unsigned long int digit = *n;
     printf("Number before flipping: %lu\n", digit);
     *n = *n ^ mask;
     printf("Number after flipping: %lu\n", digit);
